Scripts Update
- Simple Saving/Loading Function (using JSON)
- redone counting system
- hold function

+++ Assets/Scripts/Economy/EconomyMain.cs
@@ -8,21 +8,21 @@ namespace EconomyManager
     public class EconomyMain : MonoBehaviour
     {
         //Add Scores to your iterator
-        internal static int addScore(int points, int current)
+        internal static double addScore(int points, double current)
         {
             current += points;
             return current;
         }
 
         //Automatic farming of points
-        internal static int autoAddScore(int points, int current)
+        internal static double autoAddScore(int points, double current)
         {
             current += points;
             return current;
         }
 
         //For Purchasing Upgrades/Deductions
-        internal static int deductScore(int points, int current)
+        internal static double deductScore(int points, double current)
         {
             current -= points;
             return current;
